feat: schedule_page 모듈화
