+++ app/lib/src/pages/schedule_page/schedule_page.dart
@@ -1,0 +1,154 @@
+import 'package:flutter/material.dart';
+import 'package:table_calendar/table_calendar.dart';
+
+import 'package:talk_pilot/src/pages/schedule_page/widgets/event.dart';
+import 'package:talk_pilot/src/pages/schedule_page/widgets/event_input_form.dart';
+import 'package:talk_pilot/src/pages/schedule_page/widgets/event_list.dart';
+
+class SchedulePage extends StatefulWidget {
+  const SchedulePage({super.key});
+
+  @override
+  State<SchedulePage> createState() => _SchedulePageState();
+}
+
+class _SchedulePageState extends State<SchedulePage> {
+  final Map<DateTime, List<Event>> _events = {};
+  final TextEditingController _controller = TextEditingController();
+  DateTime _focusedDay = DateTime.now();
+  DateTime? _selectedDay;
+  bool _isInputVisible = false;
+  int? _editingIndex;
+  Color _selectedColor = Colors.red;
+
+  final List<Color> _colorOptions = [
+    Colors.red,
+    Colors.orange,
+    Colors.yellow,
+    Colors.green,
+    Colors.blue,
+    Colors.indigo,
+    Colors.purple,
+  ];
+
+  List<Event> _getEvents(DateTime day) => _events[day] ?? [];
+
+  void _toggleInput({int? editIndex, Event? event}) {
+    setState(() {
+      _isInputVisible = true;
+      _editingIndex = editIndex;
+      _controller.text = event?.title ?? '';
+      _selectedColor = event?.color ?? Colors.blue;
+    });
+  }
+
+  void _saveEvent() {
+    final text = _controller.text.trim();
+    if (text.isEmpty || _selectedDay == null) return;
+
+    final day = _selectedDay!;
+    final newEvent = Event(title: text, color: _selectedColor);
+
+    setState(() {
+      if (_editingIndex != null) {
+        _events[day]![_editingIndex!] = newEvent;
+      } else {
+        _events.putIfAbsent(day, () => []).add(newEvent);
+      }
+      _controller.clear();
+      _isInputVisible = false;
+      _editingIndex = null;
+    });
+  }
+
+  void _deleteEvent(int index) {
+    final day = _selectedDay!;
+    setState(() {
+      _events[day]!.removeAt(index);
+      if (_events[day]!.isEmpty) _events.remove(day);
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final selected = _selectedDay ?? _focusedDay;
+    final savedEvents = _getEvents(selected);
+
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('스케줄', style: TextStyle(color: Colors.white)),
+        backgroundColor: Colors.deepPurple,
+      ),
+      body: Column(
+        children: [
+          TableCalendar(
+            firstDay: DateTime.utc(2025, 1, 1),
+            lastDay: DateTime.utc(2035, 12, 31),
+            focusedDay: _focusedDay,
+            selectedDayPredicate: (day) => isSameDay(_selectedDay, day),
+            onDaySelected: (selectedDay, focusedDay) {
+              setState(() {
+                _selectedDay = selectedDay;
+                _focusedDay = focusedDay;
+                _isInputVisible = false;
+                _controller.clear();
+                _editingIndex = null;
+              });
+            },
+            headerStyle: const HeaderStyle(
+              titleCentered: true,
+              formatButtonVisible: false,
+            ),
+            calendarStyle: const CalendarStyle(
+              todayDecoration: BoxDecoration(
+                color: Colors.deepPurpleAccent,
+                shape: BoxShape.circle,
+              ),
+              selectedDecoration: BoxDecoration(
+                color: Colors.deepPurple,
+                shape: BoxShape.circle,
+              ),
+            ),
+          ),
+          Align(
+            alignment: Alignment.centerRight,
+            child: Padding(
+              padding: const EdgeInsets.only(right: 16.0, top: 8),
+              child: GestureDetector(
+                onTap: () => _toggleInput(),
+                child: Container(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 12,
+                    vertical: 8,
+                  ),
+                  decoration: BoxDecoration(
+                    color: Colors.deepPurple,
+                    borderRadius: BorderRadius.circular(20),
+                  ),
+                  child: const Text(
+                    '일정 생성',
+                    style: TextStyle(color: Colors.white),
+                  ),
+                ),
+              ),
+            ),
+          ),
+          if (_isInputVisible)
+            EventInputForm(
+              controller: _controller,
+              selectedColor: _selectedColor,
+              onColorChanged: (color) => setState(() => _selectedColor = color),
+              colorOptions: _colorOptions,
+              isEditing: _editingIndex != null,
+              onSave: _saveEvent,
+            ),
+          EventList(
+            events: savedEvents,
+            onEdit: (i, e) => _toggleInput(editIndex: i, event: e),
+            onDelete: _deleteEvent,
+          ),
+        ],
+      ),
+    );
+  }
+}
