+++ app/lib/src/pages/schedule_page/widgets/event.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class Event {
+  final String title;
+  final Color color;
+
+  Event({required this.title, required this.color});
+}
